Move workload flags under service
