+++ flag/service/service.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/giantswarm/cluster-apps-operator/flag/service/image"
 	"github.com/giantswarm/cluster-apps-operator/flag/service/release"
+	"github.com/giantswarm/cluster-apps-operator/flag/service/workload"
 )
 
 // Service is an intermediate data structure for command line configuration flags.
@@ -12,4 +13,5 @@ type Service struct {
 	Image      image.Image
 	Kubernetes kubernetes.Kubernetes
 	Release    release.Release
+	Workload   workload.Workload
 }
