+++ flag/service/workload/cluster/cluster.go
@@ -1,8 +1,8 @@
 package cluster
 
 import (
-	"github.com/giantswarm/cluster-apps-operator/flag/workload/cluster/calico"
-	"github.com/giantswarm/cluster-apps-operator/flag/workload/cluster/kubernetes"
+	"github.com/giantswarm/cluster-apps-operator/flag/service/workload/cluster/calico"
+	"github.com/giantswarm/cluster-apps-operator/flag/service/workload/cluster/kubernetes"
 )
 
 // Cluster is a data structure to hold cluster specific configuration flags.
