+++ flag/service/workload/workload.go
@@ -1,6 +1,6 @@
 package workload
 
-import "github.com/giantswarm/cluster-apps-operator/flag/workload/cluster"
+import "github.com/giantswarm/cluster-apps-operator/flag/service/workload/cluster"
 
 // Workload is a data structure to hold workload cluster specific configuration
 // flags.
